Merge branch main into dev/exce_handling

+++ samples/native-lib/test_hello2.c
@@ -30,12 +30,14 @@ test_hello2_wrapper(wasm_exec_env_t exec_env, uint32_t nameaddr,
     wasm_runtime_free(p);
 
     wasm_module_inst_t inst = wasm_runtime_get_module_inst(exec_env);
-    if (!wasm_runtime_validate_app_str_addr(inst, nameaddr)
-        || !wasm_runtime_validate_app_addr(inst, resultaddr, resultlen)) {
+    if (!wasm_runtime_validate_app_str_addr(inst, (uint64_t)nameaddr)
+        || !wasm_runtime_validate_app_addr(inst, (uint64_t)resultaddr,
+                                           (uint64_t)resultlen)) {
         return -1;
     }
-    const char *name = wasm_runtime_addr_app_to_native(inst, nameaddr);
-    char *result = wasm_runtime_addr_app_to_native(inst, resultaddr);
+    const char *name =
+        wasm_runtime_addr_app_to_native(inst, (uint64_t)nameaddr);
+    char *result = wasm_runtime_addr_app_to_native(inst, (uint64_t)resultaddr);
     return snprintf(result, resultlen,
                     "Hello, %s. This is %s! Your wasm_module_inst_t is %p.\n",
                     name, __func__, inst);
